# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Plataforma SaaS de entrevistas por voz com IA",
        "title": "App Recrutamento IA - API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/sessions/{id}": {
            "get": {
                "description": "Retrieves metadata about an interview session",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Gets interview session details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.InterviewSession"
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"invalid session id\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "{\"error\": \"session not found\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{id}/start": {
            "post": {
                "description": "Changes the status of an interview session to in_progress",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session"
                ],
                "summary": "Starts an interview session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"session started\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "{\"error\": \"invalid session id\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"error\": \"internal server error\"}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.InterviewSession": {
            "type": "object",
            "properties": {
                "audio_s3_key": {
                    "type": "string"
                },
                "candidate_id": {
                    "type": "string"
                },
                "duration_s": {
                    "type": "integer"
                },
                "ended_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.SessionStatus"
                },
                "transcript_s3_key": {
                    "type": "string"
                }
            }
        },
        "domain.SessionStatus": {
            "type": "string",
            "enum": [
                "invited",
                "in_progress",
                "done"
            ],
            "x-enum-varnames": [
                "SessionStatusInvited",
                "SessionStatusInProgress",
                "SessionStatusDone"
            ]
        }
    }
}